Full Project Folder
This should be able to open in whatever IDE you're using I think that
might be why the file you had looked so janky.

right now i have the variable "curTextArea" set up that will hold the
focus of the currently selected textbox so use that for your selected
area.